CP_VQA/requirements.txt
cirq-core==1.3.0
joblib==1.3.2
numba==0.59.0
numpy==1.23.5
pytest==8.0.1
qiskit==0.41.0
qsimcirq
qiskit-aer==0.11.2
qulacs==0.6.3
scipy==1.12.0
sympy==1.12

